Feature/official account (#15)
* feat(work) - add export,batch,linkCorp client

* fix(work) - remove 'Client' from tail component

* fix(work) - remove 'Client' from tail component

* update

* fix(contacts): fix params error

* feat(work) - add groupchat and moment

* fix(work)-user components name space

* fix(work)-user components name space

* fix(work)-external user components name space

* refactor(response) - uniform the response with pointer

* feat(work) - append reset components

* update

* feat(work) - add account service clients

* feat(work) - add account service clients

* update

* feat(work) - add oa with calendar and meeting

* update

* feat(work) - finish external contact webdrive

* feat(work) - external pay

* refactor(work) - broadcast request and response

* update

* update

* update

* feat(officialAccount) - media

+++ src/miniProgram/wxaCode/client.go
@@ -15,7 +15,7 @@ type Client struct {
 
 // 获取小程序二维码，适用于需要的码数量较少的业务场景
 // https://developers.weixin.qq.com/miniprogram/dev/api-backend/open-api/qr-code/wxacode.createQRCode.html
-func (comp *Client) CreateQRCode(path string, width int) (*http.Response, error) {
+func (comp *Client) CreateQRCode(path string, width int64) (*http.Response, error) {
 
 	var result string
 	var header = &response4.ResponseHeaderMedia{}
@@ -41,7 +41,7 @@ func (comp *Client) CreateQRCode(path string, width int) (*http.Response, error)
 
 // 获取小程序码，适用于需要的码数量较少的业务场景
 // https://developers.weixin.qq.com/miniprogram/dev/api-backend/open-api/qr-code/wxacode.get.html
-func (comp *Client) Get(path string, width int,
+func (comp *Client) Get(path string, width int64,
 	autoColor bool, lineColor *power.HashMap, isHyaline bool) (*http.Response, error) {
 
 	var result string
@@ -71,7 +71,7 @@ func (comp *Client) Get(path string, width int,
 // 获取小程序码，适用于需要的码数量极多的业务场景
 // https://developers.weixin.qq.com/miniprogram/dev/api-backend/open-api/qr-code/wxacode.getUnlimited.html
 func (comp *Client) GetUnlimited(
-	scene string, page string, width int,
+	scene string, page string, width int64,
 	autoColor bool, lineColor *power.HashMap, isHyaline bool) (*http.Response, error) {
 
 	var result string
